Ignore NaNs in Patches._index_data()
